functions/requirements.txt
azure-functions
azure-data-tables
azure-identity
google-genai